fix: 修复 CI 检查失败 - ESLint 和 Clippy 问题
**前端修复**：
- 移除 useCallback 中未使用的 globalConfig 依赖
- 拆分 useTheme.tsx 解决 Fast Refresh 警告（提取 theme-context.ts 和 useThemeHook.ts）
- 拆分 button.tsx 解决 Fast Refresh 警告（提取 button-variants.ts）
- 更新所有相关导入路径

**后端修复**：
- 移除 error.rs 中未使用的 ErrorContext 导入和辅助函数
- 修复 manager.rs 中重复定义的 shutdown_session_manager 函数
- 将 shutdown_session_manager 移到 test module 之前（符合 Clippy::items-after-test-module 规则）
- 运行 cargo fmt 修复所有格式问题

**架构改进**：
- 遵循 React Fast Refresh 最佳实践（组件与非组件分离）
- 保持代码模块化和可维护性
- 所有检查通过：ESLint + Clippy + Prettier + cargo fmt
