+++ src-tauri/src/commands/tool_commands/installation.rs
@@ -1,6 +1,6 @@
+use crate::commands::error::{AppError, AppResult};
 use crate::commands::tool_management::ToolRegistryState;
 use crate::commands::types::{InstallResult, ToolStatus};
-use crate::commands::error::{AppError, AppResult};
 use ::duckcoding::models::{InstallMethod, Tool};
 use ::duckcoding::services::proxy::config::apply_global_proxy;
 use ::duckcoding::services::InstallerService;
@@ -56,18 +56,20 @@ pub async fn install_tool(
     tracing::debug!(tool = %tool, method = %method, force = force, "安装工具（使用InstallerService）");
 
     // 获取工具定义
-    let tool_obj = Tool::by_id(&tool)
-        .ok_or_else(|| AppError::ToolNotFound { tool: tool.clone() })?;
+    let tool_obj =
+        Tool::by_id(&tool).ok_or_else(|| AppError::ToolNotFound { tool: tool.clone() })?;
 
     // 转换安装方法
     let install_method = match method.as_str() {
         "npm" => InstallMethod::Npm,
         "brew" => InstallMethod::Brew,
         "official" => InstallMethod::Official,
-        _ => return Err(AppError::ValidationError {
-            field: "method".to_string(),
-            reason: format!("未知的安装方法: {}", method),
-        }),
+        _ => {
+            return Err(AppError::ValidationError {
+                field: "method".to_string(),
+                reason: format!("未知的安装方法: {}", method),
+            })
+        }
     };
 
     // 使用 InstallerService 安装
