+++ src-tauri/src/core/error.rs
@@ -589,4 +589,3 @@ impl Serialize for AppError {
         }
     }
 }
-
