+++ src/hooks/theme-context.ts
@@ -1,0 +1,11 @@
+import { createContext } from 'react';
+
+export type Theme = 'light' | 'dark' | 'system';
+
+export interface ThemeContextType {
+  theme: Theme;
+  actualTheme: 'light' | 'dark'; // 实际应用的主题（考虑系统设置）
+  setTheme: (theme: Theme) => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
