+++ src/hooks/useThemeHook.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { ThemeContext, ThemeContextType } from './theme-context';
+
+export function useTheme(): ThemeContextType {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+}
